Revert "Remind customers of rentals starting today"
sale_rental_notifications this isn't yet fully specced out,
so I'm abandoning this iteration of the module.

If you are interested in the code so far, see the commit being reverted,
but please don't rely on any future iteration of this module working the
same as this iteration.

If you fork it for your own purposes I suggest you rename it so it
doesn't clash with any future revision of our module.

This reverts commit f5c9f19d1b1db34bd305cc292e0920aa3a96b6e0.