[O1193] Fix crash in onchange() method
This removes the on-change check that a newly-selected customer is still
allowed to buy all the items on the sale order's lines.
There was a conflict between different versions of the
onchange_partner_id() method (with different numbers of formal
arguments) when they were called within an onchange()
call triggered from a related field pointing at a field on
partner (specifically partner_id.ref).

So I've removed that extra check - the check will still be performed
when the user clicks 'Save' after all, and changing the customer after
adding order lines is quite unusual.